01-09-2024 Script Update inverted index.html
Updated the called script to include a new javascript file called "invertedScript.js" instead of making changes to the current script file to fit and include instructions for both the index.html code and the inverted index.html code. This will allow for changes to be made specifically to one of the sides of the website and it's functionalities while the other remains unaffected. Including the new inverted script file subsequently